% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_area_intersection_weights.R
\name{calculate_area_intersection_weights}
\alias{calculate_area_intersection_weights}
\title{Area Weighted Intersection}
\usage{
calculate_area_intersection_weights(x, y)
}
\arguments{
\item{x}{sf data.frame including one geometry column and one identifier column}

\item{y}{sf data.frame including one geometry column and one identifier column}
}
\value{
data.frame containing fraction of each feature in y that is
covered by each feature in x.
s
}
\description{
Returns the fractional percent of each
feature in y that is covered by each intersecting feature
in x. These can be used as the weights in an area-weighted
mean overlay analysis where y is the data source and area-
weighted means are being generated for x.
}
\details{
THIS WILL NOT WORK WITH SELF INTERSECTIONS!!!!!
}
\examples{
b1 = sf::st_polygon(list(rbind(c(-1,-1), c(1,-1),
                           c(1,1), c(-1,1),
                           c(-1,-1))))
b2 = b1 + 2
b3 = b1 + c(-0.2, 2)
b4 = b1 + c(2.2, 0)
b = sf::st_sfc(b1, b2, b3, b4)
a1 = b1 * 0.8
a2 = a1 + c(1, 2)
a3 = a1 + c(-1, 2)
a = sf::st_sfc(a1,a2,a3)
plot(b, border = 'red')
plot(a, border = 'green', add = TRUE)

sf::st_crs(b) <- sf::st_crs(a) <- sf::st_crs(5070)

b <- sf::st_sf(b, data.frame(idb = c(1, 2, 3, 4)))
a <- sf::st_sf(a, data.frame(ida = c(1, 2, 3)))

sf::st_agr(a) <- sf::st_agr(b) <- "constant"

a_b <- calculate_area_intersection_weights(a, b)
b_a <- calculate_area_intersection_weights(b, a)

}
