% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_utils.R
\name{get_grid_mapping}
\alias{get_grid_mapping}
\title{Get Grid Mapping}
\usage{
get_grid_mapping(x)
}
\arguments{
\item{x}{data.frame returned by ncmeta::nc_atts, open NetCDF object,
or character file path or url to be opened with RNetCDF::open.nc}
}
\description{
Get the grid mapping from a NetCDF file
}
\examples{
nc <- system.file("extdata/metdata.nc", package = "intersectr")
get_grid_mapping(nc)

nc <- system.file("extdata/daymet.nc", package = "intersectr")
get_grid_mapping(nc)

}
