% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_ncdf}
\alias{write_ncdf}
\title{Write output to NetCDF}
\usage{
write_ncdf(nc_file, int_data, variable_name, variable_units,
  variable_long_name = NULL, lats = NULL, lons = NULL)
}
\arguments{
\item{nc_file}{character path or url to a NetCDF source}

\item{int_data}{data.frame as output by execute_intersection}

\item{variable_name}{character name for data variable}

\item{variable_units}{character units for data variable}

\item{variable_long_name}{character optional long name for data variable}

\item{lats}{numeric vector giving representative latitude for each time series}

\item{lons}{numeric vector giving representative longitude for each time series}
}
\description{
Helper function to write output to NetCDF
}
