% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_area_intersection_weights_2.R
\name{calculate_area_intersection_weights}
\alias{calculate_area_intersection_weights}
\title{Area Weighted Intersection (areal implementation)}
\usage{
calculate_area_intersection_weights(x, y)
}
\arguments{
\item{x}{sf data.frame including one geometry column and one identifier column}

\item{y}{sf data.frame including one geometry column and one identifier column}
}
\value{
data.frame containing fraction of each feature in x that is
covered by each feature in y. e.g. If a feature from x is entirely within a feature in y,
w will be 1. If a feature from x is 50% in one feature for y and 50% in another, there
will be two rows, one for each x/y pair of features with w = 0.5 in each.
}
\description{
Returns the fractional percent of each
feature in x that is covered by each intersecting feature
in y. These can be used as the weights in an area-weighted
mean overlay analysis where x is the data source and area-
weighted means are being generated for y.
}
