% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cell_geometry.R
\name{create_cell_geometry}
\alias{create_cell_geometry}
\title{Create Cell Geometry}
\usage{
create_cell_geometry(x, y, prj, geom = NULL, buffer_dist = 0)
}
\arguments{
\item{x}{numeric center positions of x indices}

\item{y}{numeric center positions of y indices}

\item{prj}{character proj4 string for x and y}

\item{geom}{sf data.frame with geometry that cell geometry should cover}

\item{buffer_dist}{a distance to buffer the cell geometry in units of geom projection}
}
\description{
Creates cell geometry from vectors of x and y positions.
}
\details{
Intersection is performed with cell centers then geometry is constructed.
A buffer may be required to fully cover geometry with cells.
}
\examples{
library(RNetCDF)
library(sf)
variable_name <- "precipitation_amount"
nc <- open.nc(system.file("extdata/metdata.nc", package = "intersectr"))
nc_file <- file.inq.nc(nc)
nc_dim <- lapply(seq(0, nc_file$ndims-1), function(d) dim.inq.nc(nc, d))
nc_var <- lapply(seq(0, nc_file$nvars-1), function(v) var.inq.nc(nc, v))
var_names <- sapply(nc_var, function(v) v$name)
print(var_names)
var <- which(var_names == variable_name)
var_atts <- lapply(seq(0, nc_var[[var]]$natts - 1), function(a) att.inq.nc(nc, var - 1, a))
var_att_names <- sapply(var_atts, function(a) a$name)
coordinates <- att.get.nc(nc, var - 1, which(var_att_names == "coordinates") - 1)
print(coordinates)
# Could iomplement general handling for this... hand set for now.
# Either use coordinate variables or standard names for auxiliary coordinate variables.
x_var <- "lon"
y_var <- "lat"

grid_mapping <- att.get.nc(nc, var - 1, "grid_mapping")

if (!is.null(grid_mapping)) {
  prj <- ncdfgeom::get_prj(gm_atts)
} else {
  prj <- "+init=epsg:4326"
  warning("No grid mapping found. Assuming WGS84")
}

x <- var.get.nc(nc, x_var)
y <- var.get.nc(nc, y_var)

geom <- read_sf(system.file("shape/nc.shp", package = "sf"))

cell_geometry <- create_cell_geometry(x, y, prj, geom, 0)

plot(cell_geometry$geometry, lwd = 0.25)
plot(st_transform(geom$geometry, prj), add = TRUE)

}
