% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_intersection.R
\name{execute_intersection}
\alias{execute_intersection}
\title{Execute Intersection}
\usage{
execute_intersection(nc_file, variable_name, intersection_weights,
  cell_geometry, x_var, y_var, t_var, start_datetime = NULL,
  end_datetime = NULL)
}
\arguments{
\item{nc_file}{character path or OPeNDAP URL to NetCDF source}

\item{variable_name}{character name of variable in NetCDF source to access}

\item{intersection_weights}{data.frame as output by
\code{\link{calculate_area_intersection_weights}}}

\item{cell_geometry}{sf data.frame with col_ind and row_ind columns as output by
\code{\link{create_cell_geometry}}}

\item{x_var}{character variable with X axis data}

\item{y_var}{character variable with Y axis data}

\item{t_var}{character variable with T axis data}

\item{start_datetime}{character or POSIX character format is
\code{\link{strptime}} default e.g. "2010-10-10 10:10:10"}

\item{end_datetime}{character or POSIX}
}
\description{
Traverses many time steps of data applying
intersection weights for each time step.
}
\examples{

library(dplyr)
library(sf)

nc_file <- system.file("extdata/metdata.nc", package = "intersectr")
nc <- RNetCDF::open.nc(nc_file)
ncmeta::nc_vars(nc)
variable_name <- "precipitation_amount"
cv <- ncmeta::nc_coord_var(nc, variable_name)

x <- RNetCDF::var.get.nc(nc, cv$X, unpack = TRUE)
y <- RNetCDF::var.get.nc(nc, cv$Y, unpack = TRUE)

prj <- ncmeta::nc_gm_to_prj(ncmeta::nc_grid_mapping_atts(nc))
out_prj <- "+init=epsg:5070"

geom <- read_sf(system.file("shape/nc.shp", package = "sf")) \%>\%
  st_transform(out_prj)

cell_geometry <- create_cell_geometry(x, y, prj, geom, 0)

data_source_cells <- st_sf(select(cell_geometry, grid_ids))
target_polygons <- st_sf(select(geom, CNTY_ID))

st_agr(data_source_cells) <- "constant"
st_agr(target_polygons) <- "constant"

area_weights <- calculate_area_intersection_weights(
  data_source_cells,
  target_polygons)

intersected <- execute_intersection(nc_file, variable_name, area_weights,
                                    cell_geometry, cv$X, cv$Y, cv$T)

}
